requirements.txt
# Telegram Bot
python-telegram-bot[webhooks]==21.6

# Lark/Feishu SDK
lark-oapi==1.5.3

# GitHub
PyGithub==2.5.0

# LLM
anthropic==0.40.0
openai==1.55.0

# Async & Event System
aiohttp==3.11.0
redis==5.2.1

# Database
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0

# Config & Utils
pyyaml==6.0.2
pydantic==2.10.0
pydantic-settings==2.6.0
jinja2==3.1.4
python-dotenv==1.0.1

# Scheduling
apscheduler==3.10.4

# Logging
structlog==24.4.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
